requirements.txt
anyio==3.6.1
APScheduler==3.9.1
cachetools==5.2.0
certifi==2022.5.18.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
python-telegram-bot==20.0a1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
tornado==6.1
tzdata==2022.1
tzlocal==4.2
anyio==3.6.1
APScheduler==3.9.1
cachetools==5.2.0
certifi==2022.5.18.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
motor==3.0.0
pymongo==4.1.1
python-dateutil==2.8.2
python-telegram-bot==20.0a1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
tornado==6.1
tzdata==2022.1
tzlocal==4.2
anyio==3.6.1
APScheduler==3.9.1
cachetools==5.2.0
certifi==2022.5.18.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
motor==3.0.0
pymongo==4.1.1
python-dateutil==2.8.2
python-telegram-bot==20.0a1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
tornado==6.1
tzdata==2022.1
tzlocal==4.2
